requirements.txt
fastapi
uvicorn
python-dotenv
uuid6
sqlalchemy
asyncpg
pycountry